Refactor layout imports to use new content structure and improve organization
