+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/content/dashboard-content-skeleton.tsx
